Refactored existing Cart page
