+++ src/components/storefront/StorefrontRouter/StorefrontRouter.js
@@ -4,9 +4,10 @@ import { ChakraProvider } from '@chakra-ui/react';
 import storeTheme from 'themes/store';
 import { useShopify } from '../../../hooks/useShopify';
 import { Navbar, Footer } from '../';
-import Home from '../../../pages/Storefront/Home';
-import Product from '../../../pages/Storefront/Product';
+import Home from 'pages/Storefront/Home';
+import Product from 'pages/Storefront/Product';
 import PageNotFound from 'pages/Storefront/PageNotFound';
+import Cart from 'pages/Storefront/Cart';
 
 const StorefrontRouter = () => {
   const { createShop, createCheckout, fetchProducts, fetchCollections } = useShopify();
@@ -30,6 +31,9 @@ const StorefrontRouter = () => {
         </Route>
         <Route path={`${path}/products/:id`}>
           <Product />
+        </Route>
+        <Route path={`${path}/cart`}>
+          <Cart />
         </Route>
         <Route path={`${path}/*`}>
           <PageNotFound />
